TechEvents/requirements.txt
Flask==2.2.2
flask_mongoengine==1.0.0
WTForms==2.3.3
py-bcrypt==0.4
google-cloud==0.34.0
pyopenssl==23.2.0